moved ignore_static to source_app/decorators.py
