+++ securedrop/source_app/decorators.py
@@ -1,4 +1,4 @@
-from flask import redirect, url_for
+from flask import redirect, url_for, request
 from functools import wraps
 
 from source_app.utils import logged_in
@@ -9,5 +9,16 @@ def login_required(f):
     def decorated_function(*args, **kwargs):
         if not logged_in():
             return redirect(url_for('login'))
+        return f(*args, **kwargs)
+    return decorated_function
+
+
+def ignore_static(f):
+    """Only executes the wrapped function if we're not loading
+    a static resource."""
+    @wraps(f)
+    def decorated_function(*args, **kwargs):
+        if request.path.startswith('/static'):
+            return  # don't execute the decorated function
         return f(*args, **kwargs)
     return decorated_function
